+++ scripts/macros/2024/classFeatures/rogue/deviousStrikes.js
@@ -1,0 +1,26 @@
+import {cunningStrike} from './cunningStrike.js';
+export let deviousStrikes = {
+    name: 'Devious Strikes',
+    version: '1.3.32',
+    rules: 'modern',
+    midi: {
+        actor: [
+            {
+                pass: 'rollFinished',
+                macro: cunningStrike.midi.actor[0].macro,
+                priority: 300,
+                unique: 'cunningStrikeUse'
+            }
+        ]
+    },
+    config: [
+        {
+            value: 'uses',
+            label: 'CHRISPREMADES.Config.SimultaneousUses',
+            type: 'number',
+            default: 1,
+            category: 'homebrew',
+            homebrew: true
+        }
+    ]
+};